web: Split off internalSourceApi to its own file in internal/
